backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.17.0
google-adk==0.5.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.10.3
